requirements.txt
aiohttp>=3.7.4.post0
async-timeout>=3.0.1
attrs>=22.2.0
audio2numpy>=0.1.2
Brotli>=1.0.9
certifi>=2022.12.7
cffi>=1.15.1
chardet>=4.0.0
charset-normalizer>=2.1.1
discord>=1.7.3
discord.py>=1.7.3
dnspython>=2.2.1
ffmpeg>=1.4
idna>=3.4
multidict>=6.0.4
mutagen>=1.46.0
numpy>=1.21.6
pycparser>=2.21
pycryptodomex>=3.16.0
pymongo>=4.3.3
PyNaCl>=1.5.0
python-dotenv>=0.21.0
requests>=2.28.1
typing-extensions>=4.4.0
urllib3>=1.26.13
websockets>=10.4
yarl>=1.8.2
youtube-dl
yt-dlp
